+++ HISTORY.md
@@ -1,14 +1,22 @@
 History
 =======
 
+## 0.0.3
+
+* Switch to manual plugin injection for AMD (matches Node.js method).
+  [*[@ryan-roemer][]*]
+* Add fallback to "style" property for `$css`. [*[@ryan-roemer][]*]
+* Add `contains` support for `$css` assert. [*[@ryan-roemer][]*]
+* Add support for Node.js with JsDom. [*[@ryan-roemer][]*]
+
 ## 0.0.2
 
-* Extension of API with: `$css`, `$visible`, `$hidden`.  [*[atimb][]*]
+* Extension of API with: `$css`, `$visible`, `$hidden`. [*[@atimb][]*]
 
 ## 0.0.1
 
 * Initial release with API: `$val`, `$class`, `$attr`, `$prop`, `$html`,
-  `$text`. [*[ryan-roemer][]*]
+  `$text`. [*[@ryan-roemer][]*]
 
-[atimb]: https://github.com/atimb
-[ryan-roemer]: https://github.com/ryan-roemer
+[@atimb]: https://github.com/atimb
+[@ryan-roemer]: https://github.com/ryan-roemer